objectDetection/requirements.txt
opencv-python
streamlit
numpy
ultralytics
firebase-admin
google-genai
python-dotenv
